% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidmatrix.R
\name{centroidmatrix}
\alias{centroidmatrix}
\title{centroidmatrix
This function allows you to generate a matrix of centroids from a non-Euclidean distance matrix using the betadisper function in vegan.}
\usage{
centroidmatrix(factorname, speciesname, groupvarname, distancetype)
}
\arguments{
\item{factorname}{The name of a treatment factor file containing the treatment factors and other sources of variation and a grouping variable for calculation of centroids in the columns and observations in rows}

\item{speciesname}{The name of species file which has species in columns and observations in rows in the same order as in the factorname file and of the same length}

\item{groupvarname}{The name of the grouping variable in the factorname file for which centroids are calculated}

\item{distancetype}{The type of distance or dissimilarity matrix to be calculated from the species file using the vegdist function in vegan Default is "bray"}
}
\description{
centroidmatrix
This function allows you to generate a matrix of centroids from a non-Euclidean distance matrix using the betadisper function in vegan.
}
\examples{
centroidmatrix()
}
